refactor: unify session-ingest client around fetchSessionSnapshot
Rebase on main and adapt to use the existing session-ingest-client.ts
instead of creating a parallel module. The existing fetchSessionMessages
is now a thin wrapper over a new fetchSessionSnapshot core function that
uses generateInternalServiceToken for auth. extractLastAssistantMessage
is deferred to Phase 2 (analysis-service.ts) where it has a consumer.

- Delete redundant src/lib/session-ingest/ directory
- Add fetchSessionSnapshot (returns full snapshot, Sentry error reporting)
- Make fetchSessionMessages a wrapper preserving its User parameter
- Update plan document to reflect new architecture
